xscom: Add support for indirect XSCOM access
The PCB address argument becomes 64-bit, and represents an indirect access
when bit 0 is set to 1.

In that case, we pass the indirect address in bit 11 to 31 in the same way
as represented in the FIG tree and other documentations.

+++ include/xscom.h
@@ -18,12 +18,12 @@
  */
 
 /* Targeted SCOM access */
-extern int xscom_read(uint32_t gcid, uint32_t pcb_addr, uint64_t *val);
-extern int xscom_write(uint32_t gcid, uint32_t pcb_addr, uint64_t val);
+extern int xscom_read(uint32_t gcid, uint64_t pcb_addr, uint64_t *val);
+extern int xscom_write(uint32_t gcid, uint64_t pcb_addr, uint64_t val);
 
 /* This chip SCOM access */
-extern int xscom_readme(uint32_t pcb_addr, uint64_t *val);
-extern int xscom_writeme(uint32_t pcb_addr, uint64_t val);
+extern int xscom_readme(uint64_t pcb_addr, uint64_t *val);
+extern int xscom_writeme(uint64_t pcb_addr, uint64_t val);
 extern void xscom_init(void);
 
 /*
